refactor: ibc module handlers
