+++ src/mappings/mappingHandlers.ts
@@ -182,42 +182,6 @@ export async function handleLegacyBridgeSwap(event: CosmosEvent): Promise<void> 
 
   await legacySwap.save();
 }
-export async function handleIBCTransfer(event: CosmosEvent): Promise<void> {
-  const msg = event.msg;
-  logger.info(`[handleIBCTransfer] (tx ${msg.tx.hash}): indexing message ${msg.idx + 1} / ${msg.tx.decodedTx.body.messages.length}`)
-  logger.debug(`[handleIBCTransfer] (msg.msg): ${JSON.stringify(msg.msg, null, 2)}`)
-  
-  const decodedMsg = msg.msg.decodedMsg;
-  const sourcePort = decodedMsg.sourcePort;
-  const sourceChannel = decodedMsg.sourceChannel;
-  const tokenAmount = decodedMsg.token?.amount;
-  const tokenDenom = decodedMsg.token?.denom;
-  const sender = decodedMsg.sender;
-  const receiver = decodedMsg.receiver;
-  
-  if (!sourcePort || !sourceChannel || !tokenAmount || !tokenDenom || !sender || !receiver) {
-    logger.warn(`[handleIBCTransfer] (tx ${msg.tx.hash}): cannot index message (msg.msg): ${JSON.stringify(msg.msg, null, 2)}`)
-    return 
-  }
-  
-  const id = messageId(msg);
-  const transferEntity = IbcTransfer.create({
-    id,
-    toAddress: receiver,
-    fromAddress: sender,
-    amount: {amount: tokenAmount, denom: tokenDenom},
-    denom: tokenDenom,
-    sourcePort,
-    sourceChannel,
-    eventId: `${messageId(event)}-${event.idx}`,
-    messageId: id,
-    transactionId: msg.tx.hash,
-    blockId: msg.block.block.id
-  });
-
-  await transferEntity.save();
-}
-
 async function saveCw20BalanceEvent(id: string, address: string, amount: BigInt, contract: string, event: CosmosEvent) {
   await checkBalancesAccount(address, event.block.block.header.chainId);
   const msgId = messageId(event.msg);
@@ -239,3 +203,4 @@ export * from "./primitives";
 export * from "./bank";
 export * from "./dist";
 export * from "./gov";
+export * from "./ibc";
